+++ flutter_inappwebview/example/integration_test/web_storage_manager/main.dart
@@ -1,0 +1,14 @@
+import 'package:flutter_inappwebview/flutter_inappwebview.dart';
+import 'package:flutter_test/flutter_test.dart';
+
+import '../util.dart';
+
+part 'website_data_types.dart';
+
+void main() {
+  final shouldSkip = !WebStorageManager.isClassSupported();
+
+  skippableGroup('Web Storage Manager', () {
+    websiteDataTypes();
+  }, skip: shouldSkip);
+}