+++ flutter_inappwebview/example/integration_test/webview_flutter_test.dart
@@ -14,6 +14,7 @@ import 'chrome_safari_browser/main.dart' as chrome_safari_browser_tests;
 import 'in_app_localhost_server/main.dart' as in_app_localhost_server_tests;
 import 'tracing_controller/main.dart' as tracing_controller_tests;
 import 'support_methods/main.dart' as support_methods_tests;
+import 'web_storage_manager/main.dart' as web_storage_manager_tests;
 
 void main() {
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();
@@ -47,4 +48,5 @@ void main() {
   chrome_safari_browser_tests.main();
   in_app_localhost_server_tests.main();
   support_methods_tests.main();
+  web_storage_manager_tests.main();
 }